+++ src/Moxl/Xec/Action/Message/Reactions.php
@@ -11,7 +11,7 @@ class Reactions extends Action
     protected $_to;
     protected $_muc;
     protected $_id;
-    protected $_parentId;
+    protected $_parentid;
     protected $_reactions;
 
     public function request()
@@ -19,9 +19,9 @@ class Reactions extends Action
         $this->store();
 
         if ($this->_muc) {
-            Muc::message($this->_to, false, false, $this->_id, false, false, $this->_parentid, $this->_reactions);
+            Muc::message($this->_to, false, false, $this->_id, false, null, $this->_parentid, $this->_reactions);
         } else {
-            Message::simpleMessage($this->_to, false, false, $this->_id, false, false, $this->_parentid, $this->_reactions);
+            Message::simpleMessage($this->_to, false, false, $this->_id, false, null, $this->_parentid, $this->_reactions);
         }
     }