+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Movim Changelog
 
 v0.24.2 (master)
 ---------------------------
+* Move the Message files to the SQL database
+* Add thumbhash support when uploading images in the chat
 * Fix #1164 Implement XEP-0410: MUC Self-Ping (Schrödinger's Chat) basic behavior
 * Fix #1317 Handle properly the from attribute from MAM messages
 * Set a priority of XEP-0319: Last User Interaction in Presence over XEP-0203: Delayed Delivery